Fix counting points of E(Fp)

+++ FastGoat/UserGroup/EllCurve/EC.cs
@@ -92,7 +92,7 @@ public static class EC
         return new(a1.ToZnInt(p), a2.ToZnInt(p), a3.ToZnInt(p), a4.ToZnInt(p), a5.ToZnInt(p));
     }
 
-    public static EllGroup<ZnBigInt> ToZnBigInt(this EllGroup<Rational> E, int p)
+    public static EllGroup<ZnBigInt> ToZnBigInt(this EllGroup<Rational> E, BigInteger p)
     {
         var (a1, a2, a3, a4, a5) = E.Coefs;
         return new(a1.ToZnBigInt(p), a2.ToZnBigInt(p), a3.ToZnBigInt(p), a4.ToZnBigInt(p), a5.ToZnBigInt(p));
